+++ examples/manim/config.py
@@ -4,13 +4,17 @@ import click
 import cloup
 
 
-@cloup.group(invoke_without_command=True, no_args_is_help=True)
-def cfg():
+@cloup.group(
+    'config', aliases=['conf', 'cfg'],
+    invoke_without_command=True,
+    no_args_is_help=True,
+)
+def config():
     """Manages Manim configuration files."""
     pass
 
 
-@cfg.command(no_args_is_help=True)
+@config.command(no_args_is_help=True)
 @click.option(
     "-l", "--level",
     type=click.Choice(["user", "cwd"], case_sensitive=False), default="cwd",
@@ -22,13 +26,13 @@ def write(level: str, openfile: bool) -> None:
     pass
 
 
-@cfg.command()
+@config.command()
 def show():
     """Shows current configuration."""
     pass
 
 
-@cfg.command()
+@config.command()
 @click.option("-d", "--directory", default=os.getcwd())
 def export(directory):
     pass